src/tools/largerag/requirements.txt
# LargeRAG Dependencies
# LlamaIndex-based Large-scale Literature RAG System

# 核心LlamaIndex包
llama-index>=0.9.0
llama-index-core>=0.9.0

# 嵌入模型支持
llama-index-embeddings-openai>=0.1.0
openai>=1.0.0

# LLM支持
llama-index-llms-openai>=0.1.0

# 向量数据库支持 - Chroma
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0

# 文档处理
llama-index-readers-file>=0.1.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 日志和监控
structlog>=23.0.0

# 类型提示
typing-extensions>=4.5.0

# 测试依赖（开发时使用）
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0